requirements.txt
bs4
pandas
selenium
python-telegram-bot
apscheduler
lxml
httpx
